app/src/main/java/com/example/barterbarn/Conversation.java
package com.example.barterbarn;

public class Conversation {
    private String userID;
    private String userName;

    public Conversation(){}
    public Conversation(String userID, String userName) {
        this.userID = userID;
        this.userName = userName;
    }

    public String getUserID() {
        return userID;
    }

    public String getUserName() {
        return userName;
    }

    public void setUserID(String userID) {
        this.userID = userID;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

}
